requirements.txt
llama_cpp_python==0.1.56
streamlit==1.22.0
